backend/requirements.txt
fastapi
uvicorn
websockets
openai-whisper
pydub
audioop-lts